+++ backend/apps/ifc_validation/tests/tests_schema_validation_task.py
@@ -1,27 +1,20 @@
-from django.test import TestCase
+from django.test import TransactionTestCase
 from django.contrib.auth.models import User
 
 from apps.ifc_validation_models.models import *
-from apps.ifc_validation_models.decorators import requires_django_user_context
 
 from ..tasks import schema_validation_subtask
 
-class SchemaValidationTasksTestCase(TestCase):
+class SchemaValidationTasksTestCase(TransactionTestCase):
 
-    @classmethod
-    def setUpTestData(cls):
-
-        """
-        Creates a SYSTEM user in the (in-memory) test database.
-        Runs once for the whole test case.
-        """
-
+    def set_user_context():
         user = User.objects.create(id=1, username='SYSTEM', is_active=True)
-        user.save()
+        set_user_context(user)
 
-    @requires_django_user_context
     def test_schema_validation_task_creates_passed_validation_outcome(self):
 
+        # arrange
+        SchemaValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='valid_file.ifc',
             file='valid_file.ifc', 
@@ -29,12 +22,14 @@ class SchemaValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         schema_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
+        # assert
         outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 1)
@@ -42,9 +37,10 @@ class SchemaValidationTasksTestCase(TestCase):
         self.assertEqual(outcomes.first().outcome_code, ValidationOutcome.ValidationOutcomeCode.PASSED)
         self.assertEqual(outcomes.first().observed, None)
 
-    @requires_django_user_context
     def test_schema_validation_task_creates_error_validation_outcomes(self):
 
+        # arrange
+        SchemaValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='fail-alb005-scenario01-STATION_without_position.ifc',
             file='fail-alb005-scenario01-STATION_without_position.ifc', 
@@ -52,12 +48,14 @@ class SchemaValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         schema_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
+        # assert
         outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 4)
@@ -67,9 +65,10 @@ class SchemaValidationTasksTestCase(TestCase):
             self.assertTrue('Violated by:' in outcome.observed)
             self.assertTrue('On instance:' in outcome.observed)
 
-    @requires_django_user_context
     def test_schema_validation_task_creates_error_validation_outcome(self):
 
+        # arrange
+        SchemaValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='invalid_version.ifc',
             file='invalid_version.ifc', 
@@ -77,12 +76,14 @@ class SchemaValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         schema_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
+        # assert
         outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 1)
@@ -90,9 +91,10 @@ class SchemaValidationTasksTestCase(TestCase):
         self.assertEqual(outcomes.first().outcome_code, ValidationOutcome.ValidationOutcomeCode.SCHEMA_ERROR)
         self.assertEqual(outcomes.first().observed, 'Unsupported schema: ifc99')
 
-    @requires_django_user_context
     def test_schema_validation_task_creates_error_validation_outcomes(self):
 
+        # arrange
+        SchemaValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='pass-ifc001-IFC4.ifc',
             file='pass-ifc001-IFC4.ifc', 
@@ -100,12 +102,14 @@ class SchemaValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         schema_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
+        # assert
         outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 1)
@@ -114,9 +118,10 @@ class SchemaValidationTasksTestCase(TestCase):
             self.assertEqual(outcome.outcome_code, ValidationOutcome.ValidationOutcomeCode.SCHEMA_ERROR)
             self.assertIsNotNone(outcome.instance)
 
-    @requires_django_user_context
     def test_schema_validation_task_creates_error_validation_outcomes_2(self):
 
+        # arrange
+        SchemaValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='pass_reverse_comment.ifc',
             file='pass_reverse_comment.ifc', 
@@ -124,12 +129,14 @@ class SchemaValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         schema_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
+        # assert
         outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 5)
@@ -137,9 +144,10 @@ class SchemaValidationTasksTestCase(TestCase):
             self.assertEqual(outcome.severity, ValidationOutcome.OutcomeSeverity.ERROR)
             self.assertEqual(outcome.outcome_code, ValidationOutcome.ValidationOutcomeCode.SCHEMA_ERROR)
 
-    @requires_django_user_context
     def test_schema_validation_task_creates_error_validation_outcomes_3(self):
 
+        # arrange
+        SchemaValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='fail-als015-scenario01-long_last_segment.ifc',
             file='fail-als015-scenario01-long_last_segment.ifc', 
@@ -147,12 +155,14 @@ class SchemaValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         schema_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
+        # assert
         outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 7)
@@ -160,9 +170,10 @@ class SchemaValidationTasksTestCase(TestCase):
             self.assertEqual(outcome.severity, ValidationOutcome.OutcomeSeverity.ERROR)
             self.assertEqual(outcome.outcome_code, ValidationOutcome.ValidationOutcomeCode.SCHEMA_ERROR)
 
-    @requires_django_user_context
     def test_schema_validation_task_creates_error_validation_outcomes_4(self):
 
+        # arrange
+        SchemaValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='fail-als015-scenario01-long_last_segment.ifc',
             file='fail-als015-scenario01-long_last_segment.ifc', 
@@ -170,12 +181,14 @@ class SchemaValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         schema_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
+        # assert
         outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 7)
