+++ backend/apps/ifc_validation/tests_tasks.py
@@ -1,72 +1,16 @@
-from django.test import TestCase
+from django.test import TransactionTestCase
 from django.contrib.auth.models import User
 
 from apps.ifc_validation_models.models import *
-from apps.ifc_validation_models.decorators import requires_django_user_context
 
-from .tasks import syntax_validation_subtask
 from .tasks import bsdd_validation_subtask
 
-class ValidationTasksTestCase(TestCase):
+class ValidationTasksTestCase(TransactionTestCase):
 
-    @classmethod
-    def setUpTestData(cls):
-
-        """
-        Creates a SYSTEM user in the (in-memory) test database.
-        Runs once for the whole test case.
-        """
-
+    def set_user_context():
         user = User.objects.create(id=1, username='SYSTEM', is_active=True)
-        user.save()
+        set_user_context(user)
 
-    @requires_django_user_context
-    def test_syntax_validation_task_creates_passed_validation_outcome(self):
-
-        request = ValidationRequest.objects.create(
-            file_name='valid_file.ifc',
-            file='valid_file.ifc', 
-            size=280
-        )
-        request.mark_as_initiated()
-
-        syntax_validation_subtask(
-            prev_result={'is_valid': True, 'reason': 'test'}, 
-            id=request.id, 
-            file_name=request.file_name
-        )
-
-        outcomes = ValidationOutcome.objects.all()
-        self.assertIsNotNone(outcomes)
-        self.assertEqual(len(outcomes), 1)
-        self.assertEqual(outcomes.first().severity, ValidationOutcome.OutcomeSeverity.PASSED)
-        self.assertEqual(outcomes.first().outcome_code, ValidationOutcome.ValidationOutcomeCode.PASSED)
-        self.assertEqual(outcomes.first().observed, None)
-
-    @requires_django_user_context
-    def test_syntax_validation_task_creates_error_validation_outcome(self):
-
-        request = ValidationRequest.objects.create(
-            file_name='invalid_file.ifc',
-            file='invalid_file.ifc', 
-            size=7
-        )
-        request.mark_as_initiated()
-
-        syntax_validation_subtask(
-            prev_result={'is_valid': True, 'reason': 'test'}, 
-            id=request.id, 
-            file_name=request.file_name
-        )
-
-        outcomes = ValidationOutcome.objects.all()
-        self.assertIsNotNone(outcomes)
-        self.assertEqual(len(outcomes), 1)
-        self.assertEqual(outcomes.first().severity, ValidationOutcome.OutcomeSeverity.ERROR)
-        self.assertEqual(outcomes.first().outcome_code, ValidationOutcome.ValidationOutcomeCode.SYNTAX_ERROR)
-        self.assertTrue('On line 1 column 1' in outcomes.first().observed)
-
-    @requires_django_user_context
     def test_determine_aggregate_status_for_multiple_outcomes(self):
 
         # test cases
@@ -111,6 +55,7 @@ class ValidationTasksTestCase(TestCase):
         ]
 
         # arrange
+        ValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='abc.ifc',
             file='abc.ifc', 
@@ -131,9 +76,10 @@ class ValidationTasksTestCase(TestCase):
             # assert
             self.assertEqual(final_status, test_case['output'])
 
-    @requires_django_user_context
     def test_bsdd_validation_task_creates_na_validation_outcome(self):
 
+        # arrange
+        ValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='valid_file.ifc',
             file='valid_file.ifc', 
@@ -141,20 +87,22 @@ class ValidationTasksTestCase(TestCase):
         )
         request.mark_as_initiated()
 
+        # act
         bsdd_validation_subtask(
             prev_result={'is_valid': True, 'reason': 'test'}, 
             id=request.id, 
             file_name=request.file_name
         )
 
-        outcomes = ValidationOutcome.objects.all()
+        # assert
+        outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 1)
         self.assertEqual(outcomes[0].severity, ValidationOutcome.OutcomeSeverity.NOT_APPLICABLE)
 
-    @requires_django_user_context
     def test_bsdd_validation_task_creates_na_validation_outcome_2(self):
 
+        ValidationTasksTestCase.set_user_context()
         request = ValidationRequest.objects.create(
             file_name='pass_reverse_comment.ifc',
             file='pass_reverse_comment.ifc', 
@@ -168,7 +116,7 @@ class ValidationTasksTestCase(TestCase):
             file_name=request.file_name
         )
 
-        outcomes = ValidationOutcome.objects.all()
+        outcomes = ValidationOutcome.objects.filter(validation_task__request_id=request.id)
         self.assertIsNotNone(outcomes)
         self.assertEqual(len(outcomes), 1)
         self.assertEqual(outcomes[0].severity, ValidationOutcome.OutcomeSeverity.NOT_APPLICABLE)
